refactor: split up conftest
